+++ pf1.js
@@ -210,7 +210,7 @@ Hooks.once("setup", function () {
  */
 Hooks.once("ready", async function () {
   // Migrate data
-  const NEEDS_MIGRATION_VERSION = "0.77.0";
+  const NEEDS_MIGRATION_VERSION = "0.77.1";
   let PREVIOUS_MIGRATION_VERSION = game.settings.get("pf1", "systemMigrationVersion");
   if (typeof PREVIOUS_MIGRATION_VERSION === "number") {
     PREVIOUS_MIGRATION_VERSION = PREVIOUS_MIGRATION_VERSION.toString() + ".0";
@@ -455,9 +455,10 @@ Hooks.on("deleteOwnedItem", async (actor, itemData, options, userId) => {
   if (isLinkedToken) {
     let promises = [];
     if (itemData.type === "buff" && itemData.data.active) {
+      actor.effects.find((e) => e.data.origin?.indexOf(itemData._id) > 0)?.delete();
       const tokens = actor.getActiveTokens();
       for (const token of tokens) {
-        promises.push(token.toggleEffect(itemData.img));
+        promises.push(token.toggleEffect(itemData.img, { active: false }));
       }
     }
     await Promise.all(promises);